Address missing better-auth runtime env
