+++ convex/auth.ts
@@ -92,7 +92,7 @@ export const createAuthOptions = (ctx: GenericCtx<DataModel>) => {
             try {
               const actionCtx = requireActionCtx(ctx);
               await actionCtx.runMutation(
-                components.betterAuth.auth.deleteInvitation,
+                components.betterAuth.organization.deleteInvitation,
                 {
                   invitationId: data.invitation.id,
                 },
@@ -105,7 +105,7 @@ export const createAuthOptions = (ctx: GenericCtx<DataModel>) => {
             try {
               const actionCtx = requireActionCtx(ctx);
               await actionCtx.runMutation(
-                components.betterAuth.auth.deleteInvitation,
+                components.betterAuth.organization.deleteInvitation,
                 {
                   invitationId: data.invitation.id,
                 },
@@ -159,7 +159,7 @@ export const getCurrentUser = query({
 export const getUserById = query({
   args: { userId: v.string() },
   handler: async (ctx, args) => {
-    return ctx.runQuery(components.betterAuth.auth.getUser, {
+    return ctx.runQuery(components.betterAuth.user.getUser, {
       userId: args.userId,
     });
   },
@@ -215,7 +215,7 @@ export const getMemeberRole = query({
 export const checkEmailExists = query({
   args: { email: v.string() },
   handler: async (ctx, args) => {
-    return ctx.runQuery(components.betterAuth.auth.checkEmailExists, {
+    return ctx.runQuery(components.betterAuth.user.checkEmailExists, {
       email: args.email,
     });
   },
