+++ convex/betterAuth/_generated/api.ts
@@ -12,6 +12,7 @@ import type * as adapter from "../adapter.js";
 import type * as auth from "../auth.js";
 import type * as generatedSchema from "../generatedSchema.js";
 import type * as organization from "../organization.js";
+import type * as user from "../user.js";
 
 import type {
   ApiFromModules,
@@ -25,6 +26,7 @@ const fullApi: ApiFromModules<{
   auth: typeof auth;
   generatedSchema: typeof generatedSchema;
   organization: typeof organization;
+  user: typeof user;
 }> = anyApi as any;
 
 /**
