+++ convex/betterAuth/_generated/component.ts
@@ -1570,14 +1570,7 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
         Name
       >;
     };
-    auth: {
-      checkEmailExists: FunctionReference<
-        "query",
-        "internal",
-        { email: string },
-        boolean,
-        Name
-      >;
+    organization: {
       deleteInvitation: FunctionReference<
         "mutation",
         "internal",
@@ -1612,30 +1605,6 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
         },
         Name
       >;
-      getUser: FunctionReference<
-        "query",
-        "internal",
-        { userId: string },
-        null | {
-          _creationTime: number;
-          _id: string;
-          banExpires?: null | number;
-          banReason?: null | string;
-          banned?: null | boolean;
-          createdAt: number;
-          email: string;
-          emailVerified: boolean;
-          image?: null | string;
-          name: string;
-          role?: null | string;
-          timezone?: null | string;
-          updatedAt: number;
-          userId?: null | string;
-        },
-        Name
-      >;
-    };
-    organization: {
       getOrganization: FunctionReference<
         "query",
         "internal",
@@ -1665,6 +1634,37 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
           metadata?: null | string;
           name: string;
           slug: string;
+        },
+        Name
+      >;
+    };
+    user: {
+      checkEmailExists: FunctionReference<
+        "query",
+        "internal",
+        { email: string },
+        boolean,
+        Name
+      >;
+      getUser: FunctionReference<
+        "query",
+        "internal",
+        { userId: string },
+        null | {
+          _creationTime: number;
+          _id: string;
+          banExpires?: null | number;
+          banReason?: null | string;
+          banned?: null | boolean;
+          createdAt: number;
+          email: string;
+          emailVerified: boolean;
+          image?: null | string;
+          name: string;
+          role?: null | string;
+          timezone?: null | string;
+          updatedAt: number;
+          userId?: null | string;
         },
         Name
       >;
